[FIX] : 식단 중 특정 음식 삭제 접근 제한

+++ src/main/java/umc7th/bulk/mealMealItemMapping/service/command/MealMealItemMappingCommandServiceImpl.java
@@ -3,6 +3,8 @@ package umc7th.bulk.mealMealItemMapping.service.command;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import umc7th.bulk.global.error.GeneralErrorCode;
+import umc7th.bulk.global.error.exception.CustomException;
 import umc7th.bulk.meal.entity.Meal;
 import umc7th.bulk.mealItem.entity.MealItem;
 import umc7th.bulk.mealItem.repository.MealItemRepository;
@@ -11,6 +13,7 @@ import umc7th.bulk.mealMealItemMapping.entity.MealMealItemMapping;
 import umc7th.bulk.mealMealItemMapping.exception.MealMealItemMappingErrorCode;
 import umc7th.bulk.mealMealItemMapping.exception.MealMealItemMappingErrorException;
 import umc7th.bulk.mealMealItemMapping.repository.MealMealItemMappingRepository;
+import umc7th.bulk.user.domain.User;
 
 import java.util.Optional;
 
@@ -23,11 +26,14 @@ public class MealMealItemMappingCommandServiceImpl implements MealMealItemMappin
     private final MealItemRepository mealItemRepository;
 
     @Override
-    public void deleteMealMealItemMapping(Long mappingId) {
+    public void deleteMealMealItemMapping(User user, Long mappingId) {
 
         MealMealItemMapping mealMealItemMapping = mappingRepository.findById(mappingId)
                 .orElseThrow(() -> new MealMealItemMappingErrorException(MealMealItemMappingErrorCode.NOT_FOUND));
 
+        if (!mealMealItemMapping.getMeal().getDailyMeal().getMealPlan().getUser().getId().equals(user.getId())) {
+            throw new CustomException(GeneralErrorCode.FORBIDDEN_403);
+        }
         mappingRepository.deleteById(mappingId);
     }
